requirements.txt
matplotlib==3.8.2
numpy==1.23.5
pandas==2.2.3
pyro_ppl==1.8.6
PyYAML==6.0.1
PyYAML==6.0.2
scikit_learn==1.4.1.post1
scipy==1.15.2
tensorflow==2.12.0
tf_nightly==2.19.0.dev20241219
torch==2.1.1
tqdm==4.67.1
tensorflow==2.15.0
Dataloader==2.0

